requirements.txt
beautifulsoup4==4.12.3
bs4==0.0.2
EbookLib==0.18
lxml==5.3.0
python-dotenv==1.0.1
six==1.17.0
soupsieve==2.6
